Fixed parameter types in DAO methods for consistency and clarity

+++ app/src/main/java/com/example/fitnesstrackerapp/data/dao/WorkoutSetDao.kt
@@ -383,5 +383,5 @@ interface WorkoutSetDao {
         )
     """,
     )
-    suspend fun isPersonalRecord(exerciseId: Long, weight: Float, repetitions: Int): Boolean
+    suspend fun isPersonalRecord(exerciseId: Long, weight: Double, repetitions: Int): Boolean
 }
